feat: add exec subcommand and refactor run subcommand
Refactor aims to extract common behavior between exec and run.
It mostly moves out the run::Config into the executor module, as well
as moving the run_environment module to the root of the repo.

The main aim is to offload the overloaded run module that has outgrown
its initial purpose of a clap subcommand.